[FrameworkBundle] Introduce AbstractController, replacing ControllerTrait
